Añadiendo estilos

+++ src/components/TaskCard.jsx
@@ -2,16 +2,19 @@ import { useContext } from "react";
 import { TaskContext } from "../context/TaskContext";
 
 function TaskCard(props) {
-  const task= props.task;
-  const state= useContext(TaskContext);
+  const task = props.task;
+  const state = useContext(TaskContext);
   return (
-    <div>
-      <h1>{task.title}</h1>
-      <p>{task.description}</p>
+    <div className="bg-gray-800 text-white p-4 rounded-md">
+      <h1 className="text-xl font-bold capitalize">{task.title}</h1>
+      <p className="text-gray-500 text-sm">{task.description}</p>
       <button
+        className="bg-red-500 px-2 py-1 rounded-md mt-2 hover:bg-red-400"
         onClick={(e) => {
           e.preventDefault();
-          state.deleteTask(task.id);
+          if (confirm("¿En verdad Desea eliminar la tarea?")) {
+            state.deleteTask(task.id);
+          }
         }}
       >
         Delete task
